+++ docs/news.rst
@@ -19,6 +19,8 @@ Psycopg 3.1.3
   for non-ascii attribute names (:ticket:`#386`).
 - Fix handling of queries with escaped percent signs (``%%``) in `ClientCursor`
   (:ticket:`#399`).
+- Fix possible duplicated BEGIN statements emitted in pipeline mode
+  (:ticket:`#401`).
 
 
 Psycopg 3.1.2